+++ src/modals/ApproveTransactionModal/index.js
@@ -1,15 +1,13 @@
 import React from 'react';
 import { modalClose } from '../../reducers/_modal';
 
-import { lang } from 'balance-common';
+import { capitalize, lang } from 'balance-common';
 
 import MetamaskLogo from '../../components/MetamaskLogo';
 import LedgerLogo from '../../components/LedgerLogo';
 import TrezorLogo from '../../components/TrezorLogo';
 
 import Button from '../../components/Button';
-
-import { capitalize } from '../../helpers/utilities';
 
 import { StyledApproveTransaction } from './styles';
 import { StyledParagraph, StyledActions } from '../modalStyles';
