+++ src/modals/ReceiveModal/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { lang } from 'balance-common';
+import { capitalize, lang } from 'balance-common';
 
 import Card from '../../components/Card';
 import Button from '../../components/Button';
@@ -10,8 +10,6 @@ import Button from '../../components/Button';
 import arrowUp from '../../assets/arrow-up.svg';
 
 import { modalClose } from '../../reducers/_modal';
-
-import { capitalize } from '../../helpers/utilities';
 
 import {
   StyledContainer,
